# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Skill Collection API",
    "description": "API for managing skills, user authentication, and skill verifications",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Development server"
    },
    {
      "url": "https://api.skillcollection.com/api",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ChoiceQuestion": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "description": "Question text"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 2,
            "description": "Multiple choice options"
          },
          "answer": {
            "type": "string",
            "description": "Correct answer"
          },
          "_id": {
            "type": "string",
            "description": "Question ID"
          }
        },
        "required": ["question", "options", "answer"]
      },
      "ChoiceData": {
        "type": "object",
        "properties": {
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChoiceQuestion"
            },
            "description": "Array of multiple choice questions"
          },
          "userAnswers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "User's submitted answers"
          },
          "score": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "description": "Calculated score (0-100)"
          }
        }
      },
      "LevelData": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Level data ID"
          },
          "level": {
            "type": "string",
            "enum": ["choice", "p2p_interview", "interview"],
            "description": "Verification level"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "completed", "failed"],
            "description": "Level verification status"
          },
          "verifiedBy": {
            "type": "string",
            "description": "ID of user who verified this level"
          },
          "verifiedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when level was verified"
          },
          "link": {
            "type": "string",
            "description": "Discord/Google Meet link (for p2p_interview and interview)"
          },
          "choice": {
            "$ref": "#/components/schemas/ChoiceData",
            "description": "Choice level data (multiple choice questions, answers, and score)"
          }
        },
        "required": ["level", "status"]
      },
      "Verification": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Verification ID"
          },
          "userId": {
            "type": "string",
            "description": "User ID"
          },
          "skillId": {
            "type": "string",
            "description": "Skill ID"
          },
          "levelData": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LevelData"
            },
            "description": "Array of verification level attempts"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["userId", "skillId", "levelData"]
      },
...
    }
  },
  "paths": {
...
    "/verifications": {
      "post": {
        "summary": "Create new verification",
        "operationId": "createVerification",
        "tags": ["Verifications"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "skillId": {
                    "type": "string"
                  },
                  "levelData": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/LevelData"
                    }
                  }
                },
                "required": ["userId", "skillId"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Verification created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Verification"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "summary": "Get all verifications",
        "operationId": "getAllVerifications",
        "tags": ["Verifications"],
        "responses": {
          "200": {
            "description": "List of all verifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Verification"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/verifications/{id}": {
      "get": {
        "summary": "Get verification by ID",
        "operationId": "getVerificationById",
        "tags": ["Verifications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Verification"
                }
              }
            }
          },
          "404": {
            "description": "Verification not found"
          }
        }
      },
      "patch": {
        "summary": "Update verification",
        "operationId": "updateVerification",
        "tags": ["Verifications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "levelData": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/LevelData"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Verification"
                }
              }
            }
          },
          "404": {
            "description": "Verification not found"
          }
        }
      },
      "delete": {
        "summary": "Delete verification",
        "operationId": "deleteVerification",
        "tags": ["Verifications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification deleted"
          },
          "404": {
            "description": "Verification not found"
          }
        }
      }
    },
    "/verifications/{id}/submit": {
      "post": {
        "summary": "Submit level 1 answers (auto-grades, creates level 2 on pass)",
        "operationId": "submitAnswers",
        "tags": ["Verifications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Verification ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "answers": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "minItems": 1,
                    "description": "User's answers to the multiple choice questions"
                  }
                },
                "required": ["answers"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Answers submitted and graded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "passed": {
                      "type": "boolean",
                      "description": "Whether the user passed (score >= 80%)"
                    },
                    "score": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 100,
                      "description": "Calculated score"
                    },
                    "message": {
                      "type": "string",
                      "description": "Result message"
                    },
                    "verification": {
                      "$ref": "#/components/schemas/Verification"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or choice level not pending"
          },
          "404": {
            "description": "Verification not found"
          }
        }
      }
    },
    "/verifications/{id}/retry": {
      "post": {
        "summary": "Retry failed level 1 with fresh questions",
        "operationId": "retryChoice",
        "tags": ["Verifications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Verification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Level 1 reset with fresh questions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Verification"
                }
              }
            }
          },
          "400": {
            "description": "Choice level not found or not failed"
          },
          "404": {
            "description": "Verification not found"
          },
          "409": {
            "description": "Choice level is not failed"
          }
        }
      }
    },
    "/verifications/{id}/levels/{level}/complete": {
      "patch": {
        "summary": "Mark a verification level as completed",
        "operationId": "completeLevel",
        "tags": ["Verifications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Verification ID"
          },
          {
            "name": "level",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["choice", "p2p_interview", "interview"]
            },
            "description": "Verification level to complete"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "verifiedBy": {
                    "type": "string",
                    "description": "ID of the peer/interviewer who verified this level"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Level marked as completed (auto-creates next level if applicable)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Verification"
                }
              }
            }
          },
          "400": {
            "description": "Invalid level or cannot complete choice via this endpoint"
          },
          "404": {
            "description": "Verification or level not found"
          },
          "409": {
            "description": "Level is not pending or already completed"
          }
        }
      }
    }
  }
}